Do not emit identify event if key is empty; notice user otherwise (#50)
If the provided user key is empty, we do not want to emit an identify
event. We will log a warning message to the customer instead.

If a valid user has been provided though, not only should we emit the
event, but also we should add that user to the LRU cache so that we
don't unnecessarily generate future index events.