Support attribute reference lookups (#88)
Prior to this change, rule clauses could only target literal attribute
references. This commit expands that to include full reference support
(including our partial JSON pointer syntax).

The biggest change in this commit is the reliance on the 
`Either Text Bool` type. This type is used as a way to short-circuit
folding
operations within the rule evaluation engine.

- `Left Text` :: This is used to signal a failure condition which should
  halt further processing. The included `Text` value is an error
  describing the cause of the failure (e.g. invalid attribute
  reference).

- `Right Bool` :: This is used to signal that rule / clause / etc.
  processing as completed as normal and the related object either did or
  did not match.