Improve retry logic (#101)
The retry logic, as originally written, was not taking into account
several situations, as evidenced by the large number of suppression
entries we had for the SDK test harness. This commit re-writes the retry
logic to bring the behavior more inline with other SDKs and eliminate
those suppress lines.

In order to fully support the retry logic, this commit also introduces
support for the initialRetryDelay configuration option, which allows
customers to specify the base value used when determining how long to
wait before retrying.