Match package requirements to SDK (#105)
When we bumped the version numbers for the v3 release, we didn't bump
the corresponding versions in the redis store.

We didn't catch this because we don't try to build the redis package in
CI, and we certainly don't try to build it against every supported Stackage
resolver.

For now, I've taken the simplest solution to get some tests confirming
the behavior. Once I merge this forward in the v4 branch, I will look at
simplifying a lot of the redundancy that exists in our unit tests.