Drop explicit dependency on vector
I wanted compatibility with vector-0.13. To check whether any of the vector-0.13
changes were breaking, I reviewed where the existing code uses Data.Vector, and
the answer is almost nowhere, and all existing use cases can be replaced with
typeclass methods that don't require explicitly importing the module. As such,
we can get away without an explicit direct dependency on vector, which should
lighten the maintenance burden of getting the version bounds correct.

Strictly speaking it is possible to imagine a future vector upgrade breaking
this package, so we do lose something by dropping the version bounds entirely.
But since the functionality we use from the module is so basic, it's a little
hard to imagine that happening in practice.

In addition, it's possible that switching from monomorphic functions to
typeclass methods could have performance implications, but I don't think this
code is particularly perf-sensitive, and anyway inlining may yet be enough to
save us.